// src/pages/createJob/changeCorporateAccount/edit/constants/types.ts
export type TSubAddress = {
  addressNo: string;
  building?: string;
  floor?: string;
  mooNo?: string;
  soi?: string;
  road?: string;
  tambon: string;
  amphoe: string;
  province: string;
  postalCode: string;
  country: string;
  //type?: number;
};

export type TFinancialInfo = {
  registeredCapital?: number;
  revenuePerYear?: number;
  netProFitLoss?: number;
  shareholderEquity?: number;
};

export type TRegisterBusinessAddress = {
  address: TSubAddress[];
  emailAddress?: string;
  telephone?: string;
};
export type TPlaceIncorporateAddress = {
  address: TSubAddress[];
  emailAddress?: string;
  telephone?: string;
};

export type TCorporateInfo = {
  corporateCode?: string;
  name: string;
  registrationNo: string;
  taxId: string;
  dateofincorporation: string;
  isRegisteredThailand: boolean;
  isPrimaryCountry: boolean;
  registered: string;
  isRegisteredOther?: boolean;
  primary: string;
  isPrimaryOther?: boolean;
  registeredBusiness: TRegisterBusinessAddress;
  placeofIncorporation: TPlaceIncorporateAddress;
  // financial?: TFinancialInfo;
  registeredCapital?: number;
  revenuePerYear?: number;
  netProFitLoss?: number;
  shareholderEquity?: number;
};

export type TCorporateTypeAndIncome = {
  juristicType: string;
  juristicThai?: string;
  juristicForeign?: string;
  juristicOthers?: string;
  businessType: string;
  sourceOfIncome: string[];
  countrySourceOfIncome: string;
  investmentObjective: string;
};

export type TRegisteredCountryPrimaryCountryOperation = {
  registered: string;
  primary: string;
  isRegisteredThailand: boolean;
  isPrimaryCountry: boolean;
  isRegisteredOther?: boolean;
  isPrimaryOther?: boolean;
};